Making navbar as independent component
