+++ src/assets/components/Fragments/FormLogin.jsx
@@ -22,7 +22,7 @@ const FormLogin = () => {
     login(data, (status, res) => {
       if(status){
         localStorage.setItem('token', res) // ambil token dari res
-        window.location.href = "/product"
+        window.location.href = "/products"
       } else {
         setLoginFailed(res.response.data)
       }
