+++ src/main.jsx
@@ -12,6 +12,7 @@ import { Provider } from "react-redux";
 import store from "./redux/store";
 
 const router = createBrowserRouter([
+  
   {
     path: "/",
     element: <div>Hello Word</div>,
